requirements.txt
streamlit
pdfplumber
scikit-learn
spacy
pandas
numpy
python-docx